fix: keyframes interpolation at low speed

+++ src/main/java/gg/alexandre/replay/replay/ReplayPlayer.java
@@ -621,7 +621,7 @@ public class ReplayPlayer extends TickingSystem<EntityStore> {
             state.edit.fov = 1.0;
 
             for (BaseProperty<?> property : state.timeline.getProperties().values()) {
-                property.handle(state, (int) state.targetTick);
+                property.handle(state, state.targetTick);
             }
         } else {
             playerRef.getPacketHandler().writeNoCache(new SetMovementStates(new SavedMovementStates(true)));
